work

+++ main/src/ca/uwaterloo/flix/language/ast/SourcePosition.scala
@@ -86,7 +86,7 @@ final class SourcePosition private(val source: Source, val lineOneIndexed: Int, 
 
   def lineZeroIndexed: Int = lineOneIndexed - 1
 
-  def colZeroIndexed: Int = colOneIndexed - 1
+  def colZeroIndexed: Short = (colOneIndexed - 1).toShort
 
   /** Returns `true` if `this` and `o` represent the same source position. */
   override def equals(o: Any): Boolean = o match {
